+++ include/neuralnetwork/neuralnetwork.h
@@ -68,6 +68,7 @@ public:
 
   double get_percent_complete() const noexcept;
   bool has_training_data() const;
+  void create_bptt_batches(const std::vector<std::vector<double>>& inputs, const std::vector<std::vector<double>>& outputs, std::vector<std::vector<double>>& bptt_inputs, std::vector<std::vector<double>>& bptt_outputs) const;
 
   inline NeuralNetworkOptions& options() noexcept { 
     MYODDWEB_PROFILE_FUNCTION("NeuralNetwork");
@@ -98,7 +99,6 @@ private:
     std::vector<HiddenStates>& hidden_states,
     bool is_training) const;
 
-  void create_bptt_batches(const std::vector<std::vector<double>>& inputs, const std::vector<std::vector<double>>& outputs, std::vector<std::vector<double>>& bptt_inputs, std::vector<std::vector<double>>& bptt_outputs) const;
   void recreate_batch_from_indexes(NeuralNetworkHelper& neural_network_helper, const std::vector<std::vector<double>>& training_inputs, const std::vector<std::vector<double>>& training_outputs, std::vector<std::vector<double>>& shuffled_training_inputs, std::vector<std::vector<double>>& shuffled_training_outputs) const;
   void create_batch_from_indexes(const std::vector<size_t>& indexes, const std::vector<std::vector<double>>& training_inputs, const std::vector<std::vector<double>>& training_outputs, std::vector<std::vector<double>>& training_inputs_data, std::vector<std::vector<double>>& training_outputs_data) const;
   void break_indexes(const std::vector<size_t>& indexes, bool data_is_unique, std::vector<size_t>& training_indexes, std::vector<size_t>& checking_indexes, std::vector<size_t>& final_check_indexes) const;